Remove longs, use uint16. Also, stub a bunch of temp sensors
There's too much wrong with the temp sense code for this commit, so I stubbed it
The size of a long is undefined in C, so when we're making our arrays for the
groundstation we need to use integers of known size

+++ ex2_hal/athena/equipment_handler/include/tmp117.h
@@ -13,6 +13,6 @@
 
 #include <stdint.h>
 
-int tmp117_read(uint8_t sadd, float *val);
+int tmp117_read(uint8_t sadd, int16_t *val);
 
 int tmp421_init(uint8_t sadd);
